#956: clean up some registry references to fix circular references
