+++ src/blocks/effects/index.ts
@@ -15,23 +15,50 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export * from "./logger";
-export * from "./notification";
-export * from "./redirectPage";
-export * from "./clipboard";
-export * from "./forms";
-export * from "./tabs";
-export * from "./highlight";
-export * from "./vue";
-export * from "./event";
-export * from "./wait";
-export * from "./sound";
-export * from "./alert";
-export * from "./pageState";
-export * from "./hide";
-export * from "./exportCsv";
-export * from "./sidebar";
-export * from "./cancel";
-export * from "./error";
-export * from "./show";
-export * from "./telemetry";
+import { registerBlock } from "@/blocks/registry";
+import { LogEffect } from "./logger";
+import { ShowNotificationEffect } from "./notification";
+import { NavigateURLEffect, OpenURLEffect } from "./redirectPage";
+import { CopyToClipboard } from "./clipboard";
+import { FormFill, SetInputValue } from "./forms";
+import { CloseTabEffect, ActivateTabEffect } from "./tabs";
+import { HighlightEffect } from "./highlight";
+import { SetVueValues } from "./vue";
+import { ElementEvent } from "./event";
+import { WaitElementEffect } from "./wait";
+import { SoundEffect } from "./sound";
+import { AlertEffect } from "./alert";
+import { GetPageState, SetPageState } from "./pageState";
+import { HideEffect } from "./hide";
+import { ExportCsv } from "./exportCsv";
+import { HideSidebar, ShowSidebar } from "./sidebar";
+import { CancelEffect } from "./cancel";
+import { ErrorEffect } from "./error";
+import { ShowEffect } from "./show";
+import { TelemetryEffect } from "./telemetry";
+
+registerBlock(new LogEffect());
+registerBlock(new ShowNotificationEffect());
+registerBlock(new OpenURLEffect());
+registerBlock(new NavigateURLEffect());
+registerBlock(new CopyToClipboard());
+registerBlock(new FormFill());
+registerBlock(new SetInputValue());
+registerBlock(new CloseTabEffect());
+registerBlock(new ActivateTabEffect());
+registerBlock(new HighlightEffect());
+registerBlock(new SetVueValues());
+registerBlock(new ElementEvent());
+registerBlock(new WaitElementEffect());
+registerBlock(new SoundEffect());
+registerBlock(new AlertEffect());
+registerBlock(new GetPageState());
+registerBlock(new SetPageState());
+registerBlock(new HideEffect());
+registerBlock(new ExportCsv());
+registerBlock(new HideSidebar());
+registerBlock(new ShowSidebar());
+registerBlock(new CancelEffect());
+registerBlock(new ErrorEffect());
+registerBlock(new ShowEffect());
+registerBlock(new TelemetryEffect());
