+++ src/blocks/transformers/detect.ts
@@ -17,7 +17,6 @@
 
 import { Transformer } from "@/types";
 import { BlockArg, Schema } from "@/core";
-import { registerBlock } from "@/blocks/registry";
 import { propertiesToSchema } from "@/validators/generic";
 
 export class DetectElement extends Transformer {
@@ -53,6 +52,7 @@ export class DetectElement extends Transformer {
   };
 
   async transform({ selector }: BlockArg): Promise<Record<string, unknown>> {
+    // eslint-disable-next-line unicorn/no-array-callback-reference -- false positive for JQuery
     const $result = $(document).find(selector);
     return {
       count: $result.length,
@@ -60,5 +60,3 @@ export class DetectElement extends Transformer {
     };
   }
 }
-
-registerBlock(new DetectElement());
